Updated the authorization in the routes

+++ routes/userRoutes.js
@@ -1,12 +1,20 @@
-const {UserModel} = require('../models/Users.model')
+const { UserModel } = require('../models/Users.model')
 const express = require('express')
 const userRouter = express.Router()
+const jwt = require('jsonwebtoken')
 
 userRouter.get("/", async (req, res) => {
     let query = req.query
     try {
-        const users = await UserModel.find(query)
-        res.send([users])
+        const token = req.headers.authorization
+        jwt.verify(token,/*key :  */'Mohit Raj Singh', async (err, decoded) => {
+            if (decoded) {
+                const users = await UserModel.find(query)
+                res.send([users])
+            } else {
+                res.send({ "msg": "Some thing went wrong", "error": err.message })
+            }
+        })
     }
     catch (error) {
         res.send({ "msg": "Cannot get the users" })
@@ -17,8 +25,15 @@ userRouter.patch("/update/:id", async (req, res) => {
     const ID = req.params.id;
     const payload = req.body
     try {
-        await UserModel.findByIdAndUpdate({_id:ID},payload)
-        res.send({"msg":"Updated the User"})
+        const token = req.headers.authorization
+        jwt.verify(token,/*key :  */'Mohit Raj Singh', async (err, decoded) => {
+            if (decoded) {
+                await UserModel.findByIdAndUpdate({ _id: ID }, payload)
+                res.send({ "msg": "Updated the User" })
+            } else {
+                res.send({ "msg": "Some thing went wrong", "error": err.message })
+            }
+        })
     } catch (err) {
         res.send({ "msg": "Cannot update the user", "error": err.message })
     }
@@ -28,11 +43,18 @@ userRouter.delete("/delete/:id", async (req, res) => {
     const ID = req.params.id;
     const payload = req.body
     try {
-        await UserModel.findByIdAndDelete({_id:ID},payload)
-        res.send({"msg":"Deleted the User"})
+        const token = req.headers.authorization
+        jwt.verify(token,/*key :  */'Mohit Raj Singh', async (err, decoded) => {
+            if (decoded) {
+                await UserModel.findByIdAndDelete({ _id: ID }, payload)
+                res.send({ "msg": "Deleted the User" })
+            } else {
+                res.send({ "msg": "Some thing went wrong", "error": err.message })
+            }
+        })
     } catch (err) {
         res.send({ "msg": "Cannot delete the user", "error": err.message })
     }
 })
 
-module.exports={userRouter}
+module.exports = { userRouter }
